fix: increase version of cordova-support-google-services to 1.3.1
The GoogleServicesPlugin was being applied multiple times.

An old version of the cordova-support-google-services plugin (1.2.1) was being used.
This version always applies the GoogleServicesPlugin plugin.

More recent versions apply it only if it has not been applied before.
The version of this plugin must be updated to at least 1.3.0, where this fix was introduced.

References https://outsystemsrd.atlassian.net/browse/RNMT-2972