fix: revert version of google-services to 4.0.2
Updating the version of the cordova-support-google-services
plugin caused an issue in MABS 4. It was using version 4.2.0
of the google-services plugin. This version has issues when the
plugin is applied in the top-level project, which is exactly what
happens with cordova-android 6.4.0 and therefore MABS 4.

To fix this we needed to change the version. To achieve this we
either fork the plugin or add the only relevant file of the plugin
to the project. This commit implements the latter approach.

References https://outsystemsrd.atlassian.net/browse/RNMT-2972